Upgrade versions and change chant id data type to bigint/long
- Spring Boot 2.3.6
- Java 11
- telegrambots 5.1.1
- Graddle 6.8.3

+++ src/main/java/com/travelBill/api/core/user/User.java
@@ -29,7 +29,7 @@ public class User {
     )
     private List<Event> events;
 
-    private Integer telegramId;
+    private Long telegramId;
 
     @ManyToOne
     @JoinColumn(name = "current_event_id")
@@ -97,11 +97,11 @@ public class User {
         this.events = events;
     }
 
-    public Integer getTelegramId() {
+    public Long getTelegramId() {
         return telegramId;
     }
 
-    public void setTelegramId(Integer telegramId) {
+    public void setTelegramId(Long telegramId) {
         this.telegramId = telegramId;
     }
